refactor: rename old Python modules also
