+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/__init__.py
