+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/biom_wide_observation_table_writer.py
