+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/csv_wide_observation_table_writer.py
