+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/tsv_wide_observation_table_writer.py
