+++ src/taxpasta/infrastructure/application/wide_observation_table_writer/xlsx_wide_observation_table_writer.py
